New kas_wgpu::shared module (private)
